% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a zarr dataset}
\usage{
create_dataset(
  x,
  key,
  shape,
  chunk_shape,
  data_type = "float64",
  fill_value = "auto",
  missing_value = "auto",
  file_mode = "a",
  compressor = "raw",
  compression_options = list(),
  as_zarr = TRUE
)
}
\arguments{
\item{x}{A `group_handle`, `file_handle`, or `dataset_handle`}

\item{key}{The name of the dataset}

\item{shape}{The size of the data set}

\item{chunk_shape}{The size of the chunks}

\item{data_type}{The data type to store the data}

\item{fill_value}{The default value of data}

\item{missing_value}{The value to be interpreted as NA}

\item{file_mode}{How to open the file}

\item{compressor}{The compression to be used}

\item{compression_options}{The options for compression}

\item{as_zarr}{Save as zarr?}
}
\description{
Creat a dataset in zarr format on disk
}
