% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{create_file}
\alias{create_file}
\title{// the (python / h5py) I/O modes:
// r: can only read, file must exist
// r+: can read and write, file must exist
// w: can read and write, if file exists, will be overwritten
// w-: can read and write, file must not exist
// x: can read and write, file must not exist (same as w- ?!, so omitted here)
// a: can read and write}
\usage{
create_file(path, mode)
}
\description{
// the (python / h5py) I/O modes:
// r: can only read, file must exist
// r+: can read and write, file must exist
// w: can read and write, if file exists, will be overwritten
// w-: can read and write, file must not exist
// x: can read and write, file must not exist (same as w- ?!, so omitted here)
// a: can read and write
}
