% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{get_path}
\alias{get_path}
\title{Get the path of an object.}
\usage{
get_path(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Currently unused}
}
\description{
Gets the path of an object stored on disk.
}
