% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{json_to_list}
\alias{json_to_list}
\title{Convert a json string to a list}
\usage{
json_to_list(s)
}
\arguments{
\item{s}{A character vector of length 1.}
}
\value{
A list.
}
\description{
Takes a character vector and converts it to a list.
}
\examples{
zarr:::json_to_list('{"hello": "json"}')

}
