% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_to_json}
\alias{list_to_json}
\title{Convert a list to a json character vector}
\usage{
list_to_json(l)
}
\arguments{
\item{l}{A list object.}
}
\value{
A json object encoded in a character vector.
}
\description{
Takes a list and converts it to a json character vector.
}
\examples{
zarr:::list_to_json(list(hello = "json"))

}
