% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{open_dataset}
\alias{open_dataset}
\title{Open a Zarr dataset}
\usage{
open_dataset(x, key, ...)
}
\arguments{
\item{x}{the handle}

\item{key}{internal path of the data set}

\item{...}{reserved.}
}
\description{
Open a Zarr dataset from a file, group or dataset handles
}
