% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{open_zarr}
\alias{open_zarr}
\title{Open a zarr object.}
\usage{
open_zarr(path, file_mode = "a")
}
\arguments{
\item{path}{The path on disk}

\item{file_mode}{File mode to open the object}
}
\description{
Opens a zarr array saved on disk.
}
