% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readAttributesDatasetHandle}
\alias{readAttributesDatasetHandle}
\title{Read Attributes from a dataset handle}
\usage{
readAttributesDatasetHandle(d)
}
\arguments{
\item{d}{External pointer to a z5::filesystem::handle::Dataset object}
}
\value{
List with the attributes of the dataset.
}
\description{
Takes a dataset handle and returns its attributes.
}
\details{
This function uses external pointers and a wrong use can
crash your R session. Use only if you know what you are doing! For normal
use cases this function should not be necessary!
}
