% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readAttributesFileHandle}
\alias{readAttributesFileHandle}
\title{Read Attributes from a file handle}
\usage{
readAttributesFileHandle(f)
}
\arguments{
\item{f}{External pointer to a z5::filesystem::handle::File object}
}
\value{
List with the attributes of the file.
}
\description{
Takes a file handle and returns its attributes.
}
\details{
This function uses external pointers and a wrong use can
crash your R session. Use only if you know what you are doing! For normal
use cases this function should not be necessary!
}
