% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readAttributesGroupHandle}
\alias{readAttributesGroupHandle}
\title{Read Attributes from a group handle}
\usage{
readAttributesGroupHandle(g)
}
\arguments{
\item{g}{External pointer to a z5::filesystem::handle::Group object}
}
\value{
List with the attributes of the group.
}
\description{
Takes a group handle and returns its attributes.
}
\details{
This function uses external pointers and a wrong use can
crash your R session. Use only if you know what you are doing! For normal
use cases this function should not be necessary!
}
