% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Remove attributes}
\usage{
remove_attributes(x, attributes)
}
\arguments{
\item{x}{The object}

\item{attributes}{the attribute to removeAttribute}
}
\description{
Remove attributes from an object.
}
