% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{write_attributes}
\alias{write_attributes}
\title{Write attributes}
\usage{
write_attributes(x, attributes, ...)
}
\arguments{
\item{x}{The object}

\item{attributes}{A list of attributes}

\item{...}{Currently unused}
}
\description{
Write the attributes of an object or path
}
