% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr-package.R
\docType{package}
\name{zarr-package}
\alias{zarr}
\alias{zarr-package}
\title{zarr: The Zarr File Format in R}
\description{
Wrapping the z5 headers for R.
}
\details{
This package wraps C++ Code for Zarr Arrays and makes it available to R.
}
\author{
\strong{Maintainer}: Guido Kraemer \email{gkraemer@bgc-jena.mpg.de}

}
\keyword{internal}
