% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{create_dataset}
\alias{create_dataset}
\title{// the (python / h5py) I/O modes:
// r: can only read, file must exist
// r+: can read and write, file must exist
// w: can read and write, if file exists, will be overwritten
// w-: can read and write, file must not exist
// x: can read and write, file must not exist (same as w- ?!, so omitted here)
// a: can read and write
possible values for data_type:
"int8", "int16", "int32", "int64",
"uint8", "uint16", "uint32", "uint64",
"float32", "float64",
missing_value can be c("auto") or any number, should probably not be NA}
\usage{
create_dataset(
  path,
  shape,
  chunk_shape,
  data_type = "float64",
  fill_value = "auto",
  missing_value = "auto",
  file_mode = "a",
  compressor = "raw",
  compression_options = list(),
  as_zarr = TRUE
)
}
\description{
// the (python / h5py) I/O modes:
// r: can only read, file must exist
// r+: can read and write, file must exist
// w: can read and write, if file exists, will be overwritten
// w-: can read and write, file must not exist
// x: can read and write, file must not exist (same as w- ?!, so omitted here)
// a: can read and write
possible values for data_type:
"int8", "int16", "int32", "int64",
"uint8", "uint16", "uint32", "uint64",
"float32", "float64",
missing_value can be c("auto") or any number, should probably not be NA
}
