% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{read_attributes}
\alias{read_attributes}
\title{Read attributes}
\usage{
read_attributes(x)
}
\arguments{
\item{x}{The object}
}
\description{
Read the attributes of an object or path
}
\details{
This is an S3 generic to read the attributes of zarr arrays.
}
